agent-bot/requirements-dev.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.1
ruff==0.1.9
mypy==1.8.0
autoflake==2.2.1
pre-commit==3.6.0
httpx==0.25.2
respx==0.20.2
fakeredis[aioredis]==2.20.1
aiosqlite==0.19.0
types-redis==4.6.0.11
